project/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
celery==5.3.4
redis==5.0.1
mysql-connector-python==8.2.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
requests==2.31.0
openai==1.3.8
pydantic==2.5.0